Task 3 implementation of maxDigits() member function
Completed work to add maxDigits() member function.
This function compares the number of digits in this
instance to the number of digits in another instance
passed in as a parameter and returns the large of the two
number of digits.  This commit is passing all given
tests for task 3 for this assignment.