Task 5 implementation of appendDigit()
Completed adding appendDigit() to the LargeInteger
class.  This function will append the indicated digit to
become the new most significant digit of the
LargeInteger instance.  This method performs
dynamic memory allocation.  The digit array is always
maintained to have exactly enough space to hold the
current number of digits.  So if we want to append a
new one, we have to allocate a new digit array of size
numDigits + 1, copy over the current digits, append
the digit, then repoint the digit array to this new
block of memory and free up the old digit array.
This commit is passing all task 5 unit tests given
in assignment 2 for this function.